Revert "HW_1"
This reverts commit 69d000f2cb2d3864a396e929038d21a51d46d8f3.
