--- app/src/main/res/values/dimens.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <dimen name="margin_start">16dp</dimen>
-    <dimen name="margin_top">16dp</dimen>
-    <dimen name="margin_end">16dp</dimen>
-</resources>
